gradio_app/requirements.txt
gradio>=4.36.0
openai>=1.40.0
anthropic>=0.25.0
google-generativeai>=0.5.2
